+++ src/OpenCL.h
@@ -196,7 +196,7 @@ class OpenCL {
 public:
     OpenCL(int gpu, bool silent = false);
 
-    void initialize(const int channels, int batch_size = 1);
+    void initialize(const int channels, size_t batch_size = 1);
     void ensure_context_initialized(OpenCLContext & opencl_context);
     std::string get_device_name();
     bool has_fp16_compute();
@@ -209,7 +209,7 @@ private:
     void tune_sgemm();
     void process_tuners(std::string tuners);
 
-    int m_batch_size = 1;
+    size_t m_batch_size = 1;
     cl::Program m_program;
     std::string m_cl_args;
