+++ src/SMP.h
@@ -35,7 +35,7 @@
 #include <atomic>
 
 namespace SMP {
-    int get_num_cpus();
+    size_t get_num_cpus();
 
     class Mutex {
     public:
